ZXP-114 Move offer-view.service into service folder and some reformats
