+++ front-angular/src/app/services/index.ts
@@ -2,3 +2,5 @@ export * from './authentication.service';
 export * from './company.service';
 export * from './user.service';
 export * from './avis.service';
+export * from './offer-view.service';
+
